Move dialog delegates to a separate file, in the dialogs directory
